requirements.txt
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.0
pillow==10.1.0
scikit-learn==1.3.1
google-cloud-storage==2.10.0
matplotlib==3.8.0
tqdm==4.66.1
